moved analyticlinecharts style to its proper style file; changed the reference line fill color due to background color change
