+++ src/client/src/ui/analytics/components/analyticsLineChart/styled.tsx
@@ -1,9 +1,20 @@
 import { styled } from 'rt-theme'
 
+export const AnalyticsLineChartStyle = styled.div`
+  width: 100%;
+  height: 100%;
+  .recharts-cartesian-axis-ticks {
+    color: #ffffff;
+    width: 52px;
+    height: 12px;
+    opacity: 1;
+    font-size: 10px;
+  }
+`
+
 export const ToolTipStyle = styled.div`
   background-color: #14161c;
   width: 120px;
-  height: 20px;
   border-radius: 3px;
   display: flex;
   align-items: center;
@@ -12,12 +23,10 @@ export const ToolTipStyle = styled.div`
 
 export const ToolTipChildLeft = styled.div`
   width: 70px;
-  height: 12px;
   opacity: 0.6;
   font-size: 10px;
 `
 export const ToolTipChildRight = styled.div`
   width: 30px;
-  height: 12px;
   font-size: 10px;
 `
